draw: use \D'l'-style lines when no character is specified
When no character is specified, neatroff draws a line using veritcal
and horizontal bars. This can be problematic in some cases (fonts used,
little gap that looks odd in print, etc.).

It is reasonable to assume that when a line is to be printed, it should
simply be a line.